refactor(cache-schema-derive): split macro implementation modules
